investor_profiler/source_sklearn/requirements.txt
sagemaker
sagemaker_inference
pandas
pandas_gbq
google-cloud-bigquery<=1.22.0
s3fs